requirements.txt
numpy
matplotlib
filterpy
scienceplots
imageio
pillow
